refactor: use services and repositories
