+++ src/features/auth/ui/Logout.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { signOut } from '@local/features/auth/services/auth';
+import { signOut } from '@local/features/auth/services/authService';
 import { logout } from '../actions/logout';
 
 export default function Logout() {
