+++ src/features/fits/ui/Fits.tsx
@@ -1,9 +1,9 @@
-import { fetchFits } from '@local/libs/data';
+import { fetchAllFits } from '@local/features/fits/repositories/fitsRepository';
 import { FitsItem } from './FitsItem';
 import styles from './Fits.module.scss';
 
 export async function Fits() {
-  const fits = await fetchFits();
+  const fits = await fetchAllFits();
 
   return (
     <div className={styles.Fits}>
